requirements.txt
spacy
pdfplumber
python-docx
pandas
nltk
